+++ src/compiler/transformers/static-to-meta/visitor.ts
@@ -2,7 +2,7 @@ import * as d from '../../../declarations';
 import { dirname } from 'path';
 import { getModuleLegacy, resetModuleLegacy } from '../../build/compiler-ctx';
 import { parseCallExpression } from './call-expression';
-import { parseImport } from './import';
+import { parseModuleImport } from './import';
 import { parseStaticComponentMeta } from './component';
 import { parseStringLiteral } from './string-literal';
 import ts from 'typescript';
@@ -23,7 +23,7 @@ export const convertStaticToMeta = (
       if (ts.isClassDeclaration(node)) {
         return parseStaticComponentMeta(compilerCtx, typeChecker, node, moduleFile, compilerCtx.nodeMap, transformOpts);
       } else if (ts.isImportDeclaration(node)) {
-        parseImport(config, compilerCtx, buildCtx, moduleFile, dirPath, node);
+        parseModuleImport(config, compilerCtx, buildCtx, moduleFile, dirPath, node);
       } else if (ts.isCallExpression(node)) {
         parseCallExpression(moduleFile, node);
       } else if (ts.isStringLiteral(node)) {