+++ src/declarations/stencil-private.ts
@@ -15,7 +15,6 @@ import {
   Diagnostic,
   FsWatcher,
   FsWriteOptions,
-  HotModuleReplacement,
   Logger,
   LoggerTimeSpan,
   OptimizeCssInput,
@@ -257,6 +256,7 @@ export interface BuildConditionals extends Partial<BuildFeatures> {
   scriptDataOpts?: boolean;
   shadowDomShim?: boolean;
   asyncQueue?: boolean;
+  transformTagName?: boolean;
 }
 
 export type ModuleFormat = 'amd' | 'cjs' | 'es' | 'iife' | 'system' | 'umd' | 'commonjs' | 'esm' | 'module' | 'systemjs';
@@ -270,8 +270,7 @@ export interface RollupResults {
 
 export interface BuildCtx {
   buildId: number;
-  buildResults: BuildResults;
-  buildResults_next: CompilerBuildResults;
+  buildResults: CompilerBuildResults;
   buildMessages: string[];
   bundleBuildCount: number;
   collections: Collection[];
@@ -332,36 +331,6 @@ export interface BuildStyleUpdate {
 }
 
 export type BuildTask = any;
-
-export interface BuildResults {
-  buildId: number;
-  buildConditionals: {
-    shadow: boolean;
-    slot: boolean;
-    svg: boolean;
-    vdom: boolean;
-  };
-  bundleBuildCount: number;
-  components: BuildComponent[];
-  componentGraph: Map<string, string[]>;
-  diagnostics: Diagnostic[];
-  dirsAdded: string[];
-  dirsDeleted: string[];
-  duration: number;
-  entries: BuildEntry[];
-  filesAdded: string[];
-  filesChanged: string[];
-  filesDeleted: string[];
-  filesUpdated: string[];
-  filesWritten: string[];
-  hasError: boolean;
-  hasSuccessfulBuild: boolean;
-  hmr?: HotModuleReplacement;
-  hydrateAppFilePath?: string;
-  isRebuild: boolean;
-  styleBuildCount: number;
-  transpileBuildCount: number;
-}
 
 export type BuildStatus = 'pending' | 'error' | 'disabled' | 'default';
 
@@ -856,7 +825,7 @@ export interface CompilerCtx {
   hasSuccessfulBuild: boolean;
   isActivelyBuilding: boolean;
   lastComponentStyleInput: Map<string, string>;
-  lastBuildResults: BuildResults;
+  lastBuildResults: CompilerBuildResults;
   lastBuildStyles: Map<string, string>;
   moduleMap: ModuleMap;
   nodeMap: NodeMap;
@@ -1225,12 +1194,14 @@ export interface DevClientWindow extends Window {
   ['s-dev-server']: boolean;
   ['s-initial-load']: boolean;
   WebSocket: new (socketUrl: string, protos: string[]) => WebSocket;
+  devServerConfig?: DevClientConfig;
 }
 
 export interface DevClientConfig {
   basePath: string;
   editors: DevServerEditor[];
   reloadStrategy: PageReloadStrategy;
+  socketUrl?: string;
 }
 
 export interface HttpRequest {
@@ -1249,7 +1220,7 @@ export interface DevServerMessage {
   startServer?: DevServerConfig;
   serverStarted?: DevServerStartResponse;
   buildLog?: BuildLog;
-  buildResults?: BuildResults;
+  buildResults?: CompilerBuildResults;
   requestBuildResults?: boolean;
   error?: { message?: string; type?: string; stack?: any };
   isActivelyBuilding?: boolean;
@@ -1446,6 +1417,7 @@ export interface CustomElementsDefineOptions {
   exclude?: string[];
   resourcesUrl?: string;
   syncQueue?: boolean;
+  transformTagName?: (tagName: string) => string;
   jmp?: (c: Function) => any;
   raf?: (c: FrameRequestCallback) => number;
   ael?: (el: EventTarget, eventName: string, listener: EventListenerOrEventListenerObject, options: boolean | AddEventListenerOptions) => void;
@@ -1766,7 +1738,7 @@ export interface RenderNode extends HostElement {
    * Used to know the components encapsulation.
    * empty "" for shadow, "c" from scoped
    */
-  ['s-en']?: '' /*shadow*/ | 'c' /*scoped*/;
+  ['s-en']?: '' | /*shadow*/ 'c' /*scoped*/;
 }
 
 export type LazyBundlesRuntimeData = LazyBundleRuntimeData[];