+++ src/dev-server/client/events.ts
@@ -1,34 +1,32 @@
-import * as d from '../../declarations';
-
-export function emitBuildLog(win: d.DevClientWindow, buildLog: d.BuildLog) {
+export const emitBuildLog = (win: Window, buildLog: any) => {
   win.dispatchEvent(new CustomEvent(BUILD_LOG, { detail: buildLog }));
-}
+};
 
-export function emitBuildResults(win: d.DevClientWindow, buildResults: d.BuildResults) {
+export const emitBuildResults = (win: Window, buildResults: any) => {
   win.dispatchEvent(new CustomEvent(BUILD_RESULTS, { detail: buildResults }));
-}
+};
 
-export function emitBuildStatus(win: d.DevClientWindow, buildStatus: d.BuildStatus) {
+export const emitBuildStatus = (win: Window, buildStatus: string) => {
   win.dispatchEvent(new CustomEvent(BUILD_STATUS, { detail: buildStatus }));
-}
+};
 
-export function onBuildLog(win: d.DevClientWindow, cb: (buildLog: d.BuildLog) => void) {
+export const onBuildLog = (win: Window, cb: (buildLog: any) => void) => {
   win.addEventListener(BUILD_LOG, (ev: any) => {
     cb(ev.detail);
   });
-}
+};
 
-export function onBuildResults(win: d.DevClientWindow, cb: (buildResults: d.BuildResults) => void) {
+export const onBuildResults = (win: Window, cb: (buildResults: any) => void) => {
   win.addEventListener(BUILD_RESULTS, (ev: any) => {
     cb(ev.detail);
   });
-}
+};
 
-export function onBuildStatus(win: d.DevClientWindow, cb: (buildStatus: d.BuildStatus) => void) {
+export const onBuildStatus = (win: Window, cb: (buildStatus: string) => void) => {
   win.addEventListener(BUILD_STATUS, (ev: any) => {
     cb(ev.detail);
   });
-}
+};
 
 const BUILD_LOG = `devserver:buildlog`;
 const BUILD_RESULTS = `devserver:buildresults`;